Product Length Prediction
In this hackathon, the goal is to develop a machine learning model that can predict the length dimension of a product. The length of a product is an essential attribute that customers use to assess the product size before purchasing, and it is also crucial for packaging and storing products efficiently in the warehouse.